Epic yak shave.
I just want to mix foreground and background colors programmatically,
but there's not yet any standard way of doing that, let alone even
parsing color properties coming out of CSS objects. There are a few
pre-existing modules that do this (e.g., see references in the code),
but I'm not (yet?) ready to start pulling in external dependencies.
This code is more featureful than it has to be for my current needs,
to be clear; I wanted it to be a solid basis for eventual expansion.